Comments fixes
- reset test's indent
- add test
- rename variables
- add code comments
