+++ .NET/Microsoft.Recognizers.Text.NumberWithUnit/Chinese/Extractors/ChineseNumberWithUnitExtractorConfiguration.cs
@@ -15,7 +15,7 @@ namespace Microsoft.Recognizers.Text.NumberWithUnit.Chinese
             this.BuildPrefix = NumbersWithUnitDefinitions.BuildPrefix;
             this.BuildSuffix = NumbersWithUnitDefinitions.BuildSuffix;
             this.ConnectorToken = NumbersWithUnitDefinitions.ConnectorToken;
-            this.CompoundUnitRegex = new Regex(NumbersWithUnitDefinitions.CompoundUnitRegex, RegexOptions.IgnoreCase);
+            this.CompoundUnitConnectorRegex = new Regex(NumbersWithUnitDefinitions.CompoundUnitConnectorRegex, RegexOptions.IgnoreCase);
         }
          
         public abstract string ExtractType { get; }
@@ -30,7 +30,7 @@ namespace Microsoft.Recognizers.Text.NumberWithUnit.Chinese
 
         public string ConnectorToken { get; }
 
-        public Regex CompoundUnitRegex { get; }
+        public Regex CompoundUnitConnectorRegex { get; }
 
         public IExtractor IntegerExtractor { get; }
