+++ .NET/Microsoft.Recognizers.Text.NumberWithUnit/Constants.cs
@@ -17,6 +17,8 @@ namespace Microsoft.Recognizers.Text.NumberWithUnit
         public const string SYS_UNIT_COMPOUND = "builtin.unit.compound";
         public const string SYS_NUM = "builtin.num";
 
+        // For currencies without ISO codes, we use internal values prefixed by '_'. 
+        // These values should never be present in parse output.
         public const string FAKE_ISO_CODE_PREFIX = "_";
     }
 }
