+++ .NET/Microsoft.Recognizers.Text.NumberWithUnit/German/Extractors/GermanNumberWithUnitExtractorConfiguration.cs
@@ -15,7 +15,7 @@ namespace Microsoft.Recognizers.Text.NumberWithUnit.German
             this.BuildPrefix = NumbersWithUnitDefinitions.BuildPrefix;
             this.BuildSuffix = NumbersWithUnitDefinitions.BuildSuffix;
             this.ConnectorToken = string.Empty;
-            this.CompoundUnitRegex = new Regex(NumbersWithUnitDefinitions.CompoundUnitRegex, RegexOptions.IgnoreCase);
+            this.CompoundUnitConnectorRegex = new Regex(NumbersWithUnitDefinitions.CompoundUnitConnectorRegex, RegexOptions.IgnoreCase);
         }
 
         public abstract string ExtractType { get; }
@@ -30,7 +30,7 @@ namespace Microsoft.Recognizers.Text.NumberWithUnit.German
 
         public string ConnectorToken { get; }
 
-        public Regex CompoundUnitRegex { get; set; }
+        public Regex CompoundUnitConnectorRegex { get; set; }
 
         public abstract ImmutableDictionary<string, string> SuffixList { get; }
