+++ JavaScript/packages/recognizers-number-with-unit/src/resources/baseCurrency.ts
@@ -7,6 +7,6 @@
 // ------------------------------------------------------------------------------
 
 export namespace BaseCurrency {
-	export const CurrencyFractionList: ReadonlyMap<string, string> = new Map<string, string>([["CNY", "Fen|Jiao"],["__D", "Cent"],["RUB", "Kopek"],["AFN", "Pul"],["EUR", "Cent"],["ALL", "Qindarkë"],["_ALP", "Penny"],["GBP", "Penny"],["_GGP", "Penny"],["DZD", "Santeem"],["AOA", "Cêntimo"],["ARS", "Centavo"],["AMD", "Luma"],["AWG", "Cent"],["_AP", "Penny"],["SHP", "Penny"],["AUD", "Cent"],["AZN", "Qəpik"],["BSD", "Cent"],["BHD", "Fils"],["BDT", "Poisha"],["BBD", "Cent"],["BYN", "Kapyeyka"],["BZD", "Cent"],["XOF", "Centime"],["BMD", "Cent"],["BTN", "Chetrum"],["INR", "Paisa"],["BOB", "Centavo"],["USD", "Cent"],["BAM", "Fening"],["BWP", "Thebe"],["BRL", "Centavo"],["_BD", "Cent"],["BND", "Sen"],["SGD", "Cent"],["BGN", "Stotinka"],["BIF", "Centime"],["KHR", "Sen"],["XAF", "Centime"],["CAD", "Cent"],["CVE", "Centavo"],["KYD", "Cent"],["CLP", "Centavo"],["COP", "Centavo"],["KMF", "Centime"],["CDF", "Centime"],["NZD", "Cent"],["_CKD", "Cent"],["CRC", "Céntimo"],["HRK", "Lipa"],["CUC", "Centavo"],["CUP", "Centavo"],["CZK", "Haléř"],["DKK", "Øre"],["DJF", "Centime"],["DOP", "Centavo"],["EGP", "Piastre"],["ERN", "Cent"],["ETB", "Santim"],["FKP", "Penny"],["_FOK", "Oyra"],["FJD", "Cent"],["XPF", "Centime"],["GMD", "Butut"],["GEL", "Tetri"],["GHS", "Pesewa"],["GIP", "Penny"],["GTQ", "Centavo"],["GNF", "Centime"],["GYD", "Cent"],["HTG", "Centime"],["HNL", "Centavo"],["HKD", "Cent"],["HUF", "Fillér"],["ISK", "Eyrir"],["IDR", "Sen"],["IRR", "Dinar"],["IQD", "Fils"],["IMP", "Penny"],["ILS", "Agora"],["JMD", "Cent"],["JPY", "Sen"],["JEP", "Penny"],["JOD", "Piastre"],["KZT", "Tïın"],["KES", "Cent"],["_KID", "Cent"],["KPW", "Chon"],["KRW", "Jeon"],["KWD", "Fils"],["KGS", "Tyiyn"],["LAK", "Att"],["LBP", "Piastre"],["LSL", "Sente"],["ZAR", "Cent"],["LRD", "Cent"],["LYD", "Dirham"],["CHF", "Rappen"],["MOP", "Avo"],["MKD", "Deni"],["MGA", "Iraimbilanja"],["MWK", "Tambala"],["MYR", "Sen"],["MVR", "Laari"],["MRO", "Khoums"],["MUR", "Cent"],["MXN", "Centavo"],["_MD", "Cent"],["MDL", "Ban"],["MNT", "Möngö"],["MAD", "Centime"],["MZN", "Centavo"],["MMK", "Pya"],["NAD", "Cent"],["_ND", "Cent"],["NPR", "Paisa"],["NIO", "Centavo"],["NGN", "Kobo"],["_NID", "Cent"],["TRY", "Kuruş"],["NOK", "Øre"],["OMR", "Baisa"],["PKR", "Paisa"],["_PD", "Cent"],["PAB", "Centésimo"],["PGK", "Toea"],["PYG", "Céntimo"],["PEN", "Céntimo"],["_ PND", "Cent"],["PLN", "Grosz"],["QAR", "Dirham"],["RON", "Ban"],["RWF", "Centime"],["WST", "Sene"],["STD", "Cêntimo"],["SAR", "Halala"],["RSD", "Para"],["SCR", "Cent"],["SLL", "Cent"],["SBD", "Cent"],["SOS", "Cent"],["_SS", "Cent"],["_SP", "Penny"],["SSP", "Piastre"],["LKR", "Cent"],["SDG", "Piastre"],["SRD", "Cent"],["SZL", "Cent"],["SEK", "Öre"],["SYP", "Piastre"],["TWD", "Cent"],["TJS", "Diram"],["TZS", "Cent"],["THB", "Satang"],["PRB", "Kopek"],["TTD", "Cent"],["_TP", "Penny"],["TND", "Millime"],["TMT", "Tennesi"],["TVD", "Cent"],["UGX", "Cent"],["UAH", "Kopiyka"],["AED", "Fils"],["UYU", "Centésimo"],["VEF", "Céntimo"],["YER", "Fils"],["ZMW", "Ngwee"]]);
-	export const CurrencyFractionNumList: ReadonlyMap<string, number> = new Map<string, number>([["Kopek", 100],["Pul", 100],["Cent", 100],["Qindarkë", 100],["Penny", 100],["Santeem", 100],["Cêntimo", 100],["Centavo", 100],["Luma", 100],["Qəpik", 100],["Fils", 1000],["Poisha", 100],["Kapyeyka", 100],["Centime", 100],["Chetrum", 100],["Paisa", 100],["Fening", 100],["Thebe", 100],["Sen", 100],["Stotinka", 100],["Jiao", 10],["Fen", 100],["Céntimo", 100],["Lipa", 100],["Haléř", 100],["Øre", 100],["Piastre", 100],["Santim", 100],["Oyra", 100],["Butut", 100],["Tetri", 100],["Pesewa", 100],["Fillér", 100],["Eyrir", 100],["Dinar", 100],["Agora", 100],["Tïın", 100],["Chon", 100],["Jeon", 100],["Tyiyn", 100],["Att", 100],["Sente", 100],["Dirham", 1000],["Rappen", 100],["Avo", 100],["Deni", 100],["Iraimbilanja", 5],["Tambala", 100],["Laari", 100],["Khoums", 5],["Ban", 100],["Möngö", 100],["Pya", 100],["Kobo", 100],["Kuruş", 100],["Baisa", 1000],["Centésimo", 100],["Toea", 100],["Sentimo", 100],["Grosz", 100],["Sene", 100],["Halala", 100],["Para", 100],["Öre", 100],["Diram", 100],["Satang", 100],["Seniti", 100],["Millime", 1000],["Tennesi", 100],["Kopiyka", 100],["Tiyin", 100],["Hào", 10],["Ngwee", 100]]);
+	export const CurrencyFractionMapping: ReadonlyMap<string, string> = new Map<string, string>([["CNY", "Fen|Jiao"],["__D", "Cent"],["RUB", "Kopek"],["AFN", "Pul"],["EUR", "Cent"],["ALL", "Qindarkë"],["_ALP", "Penny"],["GBP", "Penny"],["_GGP", "Penny"],["DZD", "Santeem"],["AOA", "Cêntimo"],["ARS", "Centavo"],["AMD", "Luma"],["AWG", "Cent"],["_AP", "Penny"],["SHP", "Penny"],["AUD", "Cent"],["AZN", "Qəpik"],["BSD", "Cent"],["BHD", "Fils"],["BDT", "Poisha"],["BBD", "Cent"],["BYN", "Kapyeyka"],["BZD", "Cent"],["XOF", "Centime"],["BMD", "Cent"],["BTN", "Chetrum"],["INR", "Paisa"],["BOB", "Centavo"],["USD", "Cent"],["BAM", "Fening"],["BWP", "Thebe"],["BRL", "Centavo"],["_BD", "Cent"],["BND", "Sen"],["SGD", "Cent"],["BGN", "Stotinka"],["BIF", "Centime"],["KHR", "Sen"],["XAF", "Centime"],["CAD", "Cent"],["CVE", "Centavo"],["KYD", "Cent"],["CLP", "Centavo"],["COP", "Centavo"],["KMF", "Centime"],["CDF", "Centime"],["NZD", "Cent"],["_CKD", "Cent"],["CRC", "Céntimo"],["HRK", "Lipa"],["CUC", "Centavo"],["CUP", "Centavo"],["CZK", "Haléř"],["DKK", "Øre"],["DJF", "Centime"],["DOP", "Centavo"],["EGP", "Piastre"],["ERN", "Cent"],["ETB", "Santim"],["FKP", "Penny"],["_FOK", "Oyra"],["FJD", "Cent"],["XPF", "Centime"],["GMD", "Butut"],["GEL", "Tetri"],["GHS", "Pesewa"],["GIP", "Penny"],["GTQ", "Centavo"],["GNF", "Centime"],["GYD", "Cent"],["HTG", "Centime"],["HNL", "Centavo"],["HKD", "Cent"],["HUF", "Fillér"],["ISK", "Eyrir"],["IDR", "Sen"],["IRR", "Dinar"],["IQD", "Fils"],["IMP", "Penny"],["ILS", "Agora"],["JMD", "Cent"],["JPY", "Sen"],["JEP", "Penny"],["JOD", "Piastre"],["KZT", "Tïın"],["KES", "Cent"],["_KID", "Cent"],["KPW", "Chon"],["KRW", "Jeon"],["KWD", "Fils"],["KGS", "Tyiyn"],["LAK", "Att"],["LBP", "Piastre"],["LSL", "Sente"],["ZAR", "Cent"],["LRD", "Cent"],["LYD", "Dirham"],["CHF", "Rappen"],["MOP", "Avo"],["MKD", "Deni"],["MGA", "Iraimbilanja"],["MWK", "Tambala"],["MYR", "Sen"],["MVR", "Laari"],["MRO", "Khoums"],["MUR", "Cent"],["MXN", "Centavo"],["_MD", "Cent"],["MDL", "Ban"],["MNT", "Möngö"],["MAD", "Centime"],["MZN", "Centavo"],["MMK", "Pya"],["NAD", "Cent"],["_ND", "Cent"],["NPR", "Paisa"],["NIO", "Centavo"],["NGN", "Kobo"],["_NID", "Cent"],["TRY", "Kuruş"],["NOK", "Øre"],["OMR", "Baisa"],["PKR", "Paisa"],["_PD", "Cent"],["PAB", "Centésimo"],["PGK", "Toea"],["PYG", "Céntimo"],["PEN", "Céntimo"],["_ PND", "Cent"],["PLN", "Grosz"],["QAR", "Dirham"],["RON", "Ban"],["RWF", "Centime"],["WST", "Sene"],["STD", "Cêntimo"],["SAR", "Halala"],["RSD", "Para"],["SCR", "Cent"],["SLL", "Cent"],["SBD", "Cent"],["SOS", "Cent"],["_SS", "Cent"],["_SP", "Penny"],["SSP", "Piastre"],["LKR", "Cent"],["SDG", "Piastre"],["SRD", "Cent"],["SZL", "Cent"],["SEK", "Öre"],["SYP", "Piastre"],["TWD", "Cent"],["TJS", "Diram"],["TZS", "Cent"],["THB", "Satang"],["PRB", "Kopek"],["TTD", "Cent"],["_TP", "Penny"],["TND", "Millime"],["TMT", "Tennesi"],["TVD", "Cent"],["UGX", "Cent"],["UAH", "Kopiyka"],["AED", "Fils"],["UYU", "Centésimo"],["VEF", "Céntimo"],["YER", "Fils"],["ZMW", "Ngwee"]]);
+	export const CurrencyFractionalRatios: ReadonlyMap<string, number> = new Map<string, number>([["Kopek", 100],["Pul", 100],["Cent", 100],["Qindarkë", 100],["Penny", 100],["Santeem", 100],["Cêntimo", 100],["Centavo", 100],["Luma", 100],["Qəpik", 100],["Fils", 1000],["Poisha", 100],["Kapyeyka", 100],["Centime", 100],["Chetrum", 100],["Paisa", 100],["Fening", 100],["Thebe", 100],["Sen", 100],["Stotinka", 100],["Jiao", 10],["Fen", 100],["Céntimo", 100],["Lipa", 100],["Haléř", 100],["Øre", 100],["Piastre", 100],["Santim", 100],["Oyra", 100],["Butut", 100],["Tetri", 100],["Pesewa", 100],["Fillér", 100],["Eyrir", 100],["Dinar", 100],["Agora", 100],["Tïın", 100],["Chon", 100],["Jeon", 100],["Tyiyn", 100],["Att", 100],["Sente", 100],["Dirham", 1000],["Rappen", 100],["Avo", 100],["Deni", 100],["Iraimbilanja", 5],["Tambala", 100],["Laari", 100],["Khoums", 5],["Ban", 100],["Möngö", 100],["Pya", 100],["Kobo", 100],["Kuruş", 100],["Baisa", 1000],["Centésimo", 100],["Toea", 100],["Sentimo", 100],["Grosz", 100],["Sene", 100],["Halala", 100],["Para", 100],["Öre", 100],["Diram", 100],["Satang", 100],["Seniti", 100],["Millime", 1000],["Tennesi", 100],["Kopiyka", 100],["Tiyin", 100],["Hào", 10],["Ngwee", 100]]);
 }
