+++ JavaScript/packages/recognizers-number-with-unit/src/resources/frenchNumericWithUnit.ts
@@ -11,7 +11,7 @@ export namespace FrenchNumericWithUnit {
 	export const AgeSuffixList: ReadonlyMap<string, string> = new Map<string, string>([["Ans", "ans"],["Mois", "mois d'âge|mois d'age|mois"],["Semaines", "semaine|semaines|semaines d'âge|semaines d'age"],["Jour", "jours|jour"]]);
 	export const AreaSuffixList: ReadonlyMap<string, string> = new Map<string, string>([["Kilomètre carré", "km2|km^2|km²|kilomètres carrés|kilomètre carré"],["Hectomètre carré", "hm2|hm^2|hm²|hectomètre carré|hectomètres carrés"],["Décamètre carré", "dam2|dam^2|dam²|décamètre carré|décamètres carrés"],["Mètre carré", "m2|m^2|m²|mètre carré|mètres carrés"],["Décimètre carré", "dm2|dm^2|dm²|décimètre carré|décimètres carrés"],["Centimètre carré", "cm2|cm^2|cm²|centimètre carré|centimètres carrés"],["Millimètre carré", "mm2|mm^2|mm²|millimètre carré|millimètres carrés"],["Pouce carré", "pouces2|po2|pouce carré|pouces carrés|in^2|in²|in2"],["Pied carré", "pied carré|pieds carrés|pi2|pi^2|pi²"],["Mile carré", "mile carré|miles carrés|mi2|mi^2|mi²"],["Acre", "acre|acres"]]);
 	export const CurrencySuffixList: ReadonlyMap<string, string> = new Map<string, string>([["Abkhazie apsar", "abkhazie apsar|apsars"],["Afghan afghani", "afghan afghani|؋|afn|afghanis|afghani"],["Pul", "pul"],["Euro", "euros|euro|€|eur|d'euros"],["Cent", "cents|cent|-cents|-cent|sen"],["lek Albanais", "lek albanais|leks|lek"],["Qindarkë", "qindarkë|qindarkës|qindarke|qindarkes"],["Kwanza angolais", "kwanza angolais|kz|aoa|kwanza|kwanzas"],["Dram arménien", "dram arménien|drams arméniens"],["Florins d'Aruba", "florins aruba|ƒ|awg"],["Bangladeshi taka", "bangladeshi taka|৳|bdt|taka|takas|bangladeshi takas"],["Paisa", "poisha|paisa"],["Ngultrum bhoutanais", "ngultrum bhoutanais|nu.|btn"],["Chetrum", "chetrums|chetrum"],["Boliviano bolivien", "boliviano bolivien|bolivianos bolivien|bolivianos bolivie|boliviano bolivie|bob|bs."],["Bosnie-Herzégovine mark convertible", "bosnie-herzégovine mark convertible|bosnie-et-herzégovine mark convertible|bam"],["Fening", "fening|fenings"],["Pula", "pula|bwp"],["Thebe", "thebe"],["Réal brésilien", "réal brésilien|réals brésilien|r$|brl|real bresil|reals bresilien"],["Lev bulgare", "lev bulgare|levs bulgare|lv|bgn"],["Stotinki búlgaro", "stotinki bulgare"],["Riel cambodgien", "riel cambodgien|khr|៛"],["Escudo du cap-vert", "escudo cap-verdien|cve"],["Colon du costa rica", "colon du costa rica|colons du costa rica|crc|₡"],["Colon du salvador", "colon du salvador|colons du salvador|svc"],["Kuna croate", "kuna croate|kunas croate|kn|hrk"],["Lipa", "lipa"],["Couronne tchèque", "couronne tchèque|couronnes tchèque|czk|Kč"],["Haléř", "haléř"],["Nakfas érythréens", "nakfas érythréens|nfk|ern|nakfa érythréens"],["Birr éthiopien", "birr éthiopien|birrs éthiopien|etb"],["Dalasi gambienne", "gmd"],["Butut", "bututs|butut"],["Lari géorgien", "lari géorgie|lari géorgiens|gel|₾"],["Tetri géorgien", "tetri géorgie|tetris géorgiens"],["Cedi", "cedi|ghs|cedi ghanéen|gh₵"],["Pesewa", "pesewa|pesewas"],["Quetzal guatémaltèque", "quetzal guatémaltèque|gtq|quetzal|quetzales"],["Gourdes haïtiennes", "gourdes haïtiennes|gourdes|htg|gourde haïtienne"],["Lempira hondurien", "lempira hondurien|hnl"],["Forint hongrois", "forint hongrois|huf|fg|forints hongrois"],["Fillér", "fillér"],["Rial iranien", "rial iranien|irr|rials iranien|rials iraniens"],["Litas lituanien", "litas lituanien|ltl|lit lithuanien|litas lithuanie"],["Yen Japonais", "yen japonais|yen japon|yens|jpy|yen|¥|-yen"],["Tenge kazakh", "tenge kazakh|kzt"],["Shilling kényan", "shilling kényan|sh|kes|shillings kényans"],["Won coréen", "won coréen|won coréens|₩"],["Won sud-coréen", "won sud-coréen|won sud coréen|won sud-coréens|krw"],["Corée du nord won", "corée du nord won|corée nord won|kpw"],["Som Kirghizie", "som kirghizie|kgs"],["Sum Ouzbékistan", "sum ouzbékistan|sum ouzbeks|sum ouzbéks|uzs"],["Kip laotien", "kip laotien|lak|₭n|₭"],["Att", "att"],["Loti", "loti|maloti|lsl"],["Sente", "sente|lisente"],["Rand sud-africain", "rand sud-africain|zar"],["Pataca macanais", "pataca macanais|mop$|mop"],["Avo", "avos|avo"],["Dinar macédonien", "dinar macédonien|mkd|ден"],["Deni", "deni"],["Ariary malagache", "ariary malagache|mga"],["Iraimbilanja", "Iraimbilanja"],["Kwacha malawien", "kwacha malawien|mk|mwk"],["Tambala", "Tambala"],["Ringitt malaisien", "ringitt malaisien|rm|myr|ringitts malaisien"],["Ouguiya mauritanienne", "ouguiya|um|mro|ouguiya mauritanien|ouguiya mauritanienne"],["Khoums", "khoums"],["Togrogs mongoles", "togrogs mongoles|togrogs|tugriks|tögrög|mnt|₮|tögrög mongoles|tögrög mongolie|togrogs mongolie"],["Metical mozambique", "metical du mozambique|metical mozambique|mt|mzn|meticals mozambique"],["Kyat birmanie", "kyat birmanie|ks|mmk"],["Pya", "pya"],["Cordoba nicaraguayen", "cordoba nicaraguayen|córdoba nicaraguayen|nio|córdoba oro|cordoba oro nicaraguayen"],["Naira nigérians", "naira nigérians|naira|ngm|₦|nairas nigérians"],["Livre turque", "livre turque|try|tl|livre turques"],["Kuruş", "kuruş"],["Rials omanais", "rials omanais|omr|ر.ع.|rial omanais"],["Balboa panaméennes", "balboa panaméennes|balboa|pab"],["Kina", "kina|pkg|pgk"],["Toea", "toea"],["Guaraní paraguayen", "guaraní paraguayen|₲|pyg"],["Sol péruvien", "nuevo sol péruvien|soles|sol|sol péruvien"],["Złoty polonais", "złoty polonais|złoty|zł|pln|zloty|zloty polonais"],["Groxz", "groszy|grosz|grosze"],["Riyal qatari", "riyal qatari|qar|riyals qatari"],["Riayl saudi", "riyal saudi|sar|riyals saudi"],["Riyal", "riyal|riyals|rial|﷼"],["Dirham", "dirham|dirhem|dirhm"],["Halala", "halalas|halala"],["Tala", "tala|tālā|ws$|sat|wst"],["Sene", "sene"],["Dobra", "dobra|db|std"],["Leone", "leone|sll"],["Florins Néerlandais", "florins hollandais|florins néerlandais|florins|ang|ƒ|florin|fl |"],["Lilangeni", "lilangeni|szl"],["Somoni tadjikistan", "somoni tadjikistan|tjs|somoni"],["Diram", "dirams|diram"],["Baht thaïlandais", "baht thaïlandais|baht thailandais|baht thaï|baht thai|baht|฿|thb"],["Satang", "satang|satangs"],["Paʻanga", "paʻanga|pa'anga|top"],["Hryvnia ukrainien", "hryvnia ukrainien|hyrvnia|uah|₴|hryvnias ukrainien|hryvnia ukrainienne"],["Vanuatu vatu", "vanuatu vatu|vatu|vuv"],["Bolívar vénézuélien", "bolívar vénézuélien|bolivar venezuelien|bs.f.|vef|bolívars vénézuélien|bolivars venezuelien"],["Dong vietnamien", "dong vietnamien|dongs vietnamiens|dong|đồng|vnd|dông|dông vietnamiens"],["Kwacha de Zambie", "kwacha de zambie|zk|zmw|kwachas"],["Dirham marocain", "dirham marocain|mad|د.م."],["Dirham des Émirats arabes unis", "dirham des Émirats arabes unis|د.إ|aed"],["Manat azerbaïdjanais", "manat azerbaïdjanais|manat azerbaidjanais|azn"],["Manat turkmène", "manat turkmène|tmt|manat turkmene"],["Manat", "manats|manat"],["Qəpik", "qəpik"],["Shilling somalien", "shilling somalien|shillings somalien|sos"],["Shilling tanzanien", "shilling tanzanien|shillings tanzanien|tzs|tsh|shilling tanzanienne|shillings tanzanienne"],["Shilling ougandais", "shilling ougandais|shillings ougandais|sh|ugx"],["Leu roumain", "leu roumain|lei|leu roumaine|ron"],["Leu moldave", "leu meoldave|mdl"],["Leu", "leu"],["Ban", "bani|-ban|ban"],["Roupie népalaise", "roupie népalaise|roupie nepalaise|npr"],["Roupie pakistanaise", "roupie pakistanaise|pkr"],["Roupie indienne", "roupie indienne|inr|roupie indien|inr|₹"],["Roupie seychelloise", "roupie seychelloise|scr|sr|sre"],["Roupie mauricienne", "roupie mauricienne|mur"],["Rufiyaa maldives", "rufiyaa maldives|mvr|.ރ|rf"],["Roupie srilankaise", "roupie srilankaise|lrk|රු|ரூ"],["Rupiah Indonésie", "rupia indonésie|rupia indonesie|rupiah|rp|idr"],["Roupie", "roupie"],["Couronne danoise", "couronne danoise|dkk|couronnes danoise|couronne danemark|couronnes danemark"],["Couronne norvégienne", "couronne norvégienne|couronne norvegienne|couronnes norvégienne|couronnes norvegienne|nok"],["Couronne féroïenne", "couronne féroïenne|couronne feroienne"],["Couronne suédoise", "couronne suédoise|couronne suéde|sek|couronnes suédoise|couronne suedoise"],["Couronne", "couronne|couronnes"],["Øre", "Øre|oyra|eyrir"],["Franc CFA de l'Afrique de l'Ouest", "franc cfa de l''afrique de l''ouest|franc cfa ouest africain|franc cfa|francs cfa|fcfa|frs cfa|cfa francs|xof"],["Franc CFA d'Afrique centrale", "franc cfa d''afrique centrale|franc cfa centrale|frs cfa centrale|xaf"],["Franc comorien", "franc comorien|kmf"],["Franc congolais", "franc congolais|cdf"],["Franc burundais", "franc burundais|bif"],["Franc djiboutienne", "franc djiboutienne|djf"],["Franc CFP", "franc cfp|xpf"],["Franc guinéen", "franc guinéen|gnf"],["Franc Suisse", "franc suisse|chf|sfr.|francs suisses"],["Franc rwandais", "franc rwandais|rwf|rw|r₣|frw"],["Franc belge", "franc belge|bi.|b.fr.|bef"],["Rappen", "rappen|-rappen"],["Franc", "francs|franc|fr.|fs"],["Centimes", "centimes|centime|santim"],["Rouble russe", "rouble russe|rub|₽|₽ russe|roubles russe|roubles russes|₽ russes"],["Nouveau rouble biélorusse", "nouveau rouble biélorusse|byn|nouveau roubles biélorusse|nouveau rouble bielorusse|nouveau roubles biélorusse"],["Rouble transnistriens", "rouble transnistriens|prb"],["Rouble biélorusses", "rouble biélorusses|roubles biélorusses|rouble bielorusses|roubles bielorusses"],["Kopek", "kopek|kopeks"],["Kapyeyka", "kapyeyka"],["Rouble", "roubles|rouble|br"],["Dinar algérien", "dinar algérien|د.ج|dzd|dinars algérien|dinar algerien|dinars algerien"],["Dinar de bahreïn", "dinar de bahreïn|bhd|.د.ب|dinar de bahrein"],["Santeem", "santeem|santeems"],["Dinar iraquien", "dinar iraquien|dinars iraquien|iqd|ع.د|dinar iraquienne|dinars iraquienne"],["Dinar jordanien", "dinar jordanien|dinars jordanien|د.ا|jod"],["Dinar koweïtien", "dinar koweïtien|dinar koweitien|dinars koweïtien|kwd|د.ك"],["Dinar libyen", "dinar libyen|dinars libyen|lyd"],["Dinar serbe", "dinar serbe|dinars serbe|rsd|дин."],["Dinar tunisien", "dinar tunisien|dinars tunisien|tnd"],["Dinar yougoslave", "dinar yougoslave|dinars yougoslave|yun"],["Dinar", "dinars|dinar|denar|-dinars|-dinar"],["Fils", "fils|fulūs|-fils|-fil"],["Para", "para|napa"],["Millime", "millimes|millime"],["Peso argentin", "peso argentin|ars|pesos argentin|peso argentine|pesos argentine"],["Peso chilien", "peso chilien|pesos chilien|clp"],["Peso colombien", "peso colombien|pesos colombien|cop|peso colombie|pesos colombien"],["Peso cubains convertibles", "peso cubains convertibles|pesos cubains convertibles|cuc"],["Peso cubains", "peso cubaines|pesos cubaines|peso cubaine|pesos cubaines|cup"],["Peso dominicain", "peso dominicain|pesos dominicain|dop|peso dominicaine|pesos dominicaine"],["Peso philippin", "peso philippin|pesos philippin|piso|₱|php"],["Peso uruguayen", "peso uruguayen|pesos uruguayen|uyu"],["Peso", "pesos|Peso"],["Centavo", "centavos|Centavo"],["Livre britannique", "livre britannique|livres britannique|gbp|£ britannique"],["Livre guernesey", "livre guernesey|£ guernesey|ggp"],["Livre ascension", "livre ascension|livres ascension|£ ascension"],["Livre sainte-hélène", "livre de sainte-hélène|livre sainte-hélène|livre sainte-helene|livre de sainte hélène|shp"],["Livre égyptienne", "livre égyptienne|livre egyptienne|egp|ج.م"],["Livre des îles falkland", "livre des îles falkland|livre des iles falkland|fkp|£ iles falkland"],["Livre gibraltar", "livre gibraltar|livre de gibraltar|£ gibraltar|gip"],["Livre manx", "imp|livre manx|£ manx"],["Livre jersey", "livre de jersey|livre jersey|jep|£ jersey"],["Livre libanaise", "livre libanaise|£ libanaise|livres libanaise|lbp|ل.ل"],["Livre des îles malouines", "livre des îles malouines|livre des iles malouines|£ iles malouines"],["Livre sud-soudanaise", "livre sud-soudanaise|livre sud soudanaise|livre du soudan du sud|livres sud-soudanaises|livre sud soudan|livre soudan sud"],["Livre soudanaise", "livre soudanaise|livres soudanaise|sdg|£ soudan|ج.س.|livre soudan|livres soudan"],["Livre syrienne", "livre syrienne|ل.س|syp|livre syrie|livres syrie|£ syrie"],["Livre", "livre|livres|-livre|-livres|£"],["Pence", "pence"],["Shilling", "shilling|shillings"],["Penny", "penny|sou|centime"],["Dollar Américain", "dollar américain|$ américain|$ americain|usd|$usd|$ usd|dollar americain|dollar États-Unis|dollar des États-Unis|dollar États Unis|dollar etats unis|dollar etats-unis|$ etats-unis|$ États-Unis"],["Dollar des Caraïbes orientales", "dollar des caraïbes orientales|dollar des caraibes orientales|xcd|$ caraibes orientales|$ caraïbes orientales"],["Dollar Australien", "dollar australien|dollars australiens|$ australien|aud|$australien|australien $|$ australie|dollar australie"],["Dollar des bahamas", "dollar des bahamas|dollar bahamas|$ bahamas|bsd|bahama $|dollar bahama|$ bahamas"],["Dollar bermudes", "dollar des bermudes|dollar bermude|dollar bermudes|$ bermudes|bmd"],["Dollar belize", "dollar de Belize|dollar belizien|bzd|$ belize"],["Dollar îles Vierges britanniques", "dollar îles vierges britanniques|dollar iles vierges britanniques|$ iles vierges britanniques"],["Dollar brunei", "dollar de brunei|$ brunei|bnd|dollar brunei"],["Sen", "sen"],["Dollar singapour", "dollar de singapour|dollar singapour|$ sinapour|sgd|$s"],["Dollar Canadien", "dollar canadien|dollars canadien|$ canadien|cad|$can|$c|$ c|dollar canada|dollar canadienne|$ canada|$cad|cad$"],["Dollar iles caimanes", "dollars des îles caïmanes|dollar des îles caïmanes|dollars des iles caimanes|dollar iles caimanes|kyd|$ci"],["Dollar néo-zélandais", "dollar néo-zélandais|dollar néo zélandais|dollar neo-zelandais|dollar neo zelandais|$nz|$ néo-zélandais|$ neo zelandais"],["Dollar îles cook", "dollar îles cook|dollar iles cook|$ iles cook"],["Dollar des fidji", "dollar des fidji|$ fidji|dollar fidji|dollar de fidji|dollars des fidji|dollars de fidji"],["Dollar guyanien", "dollar guyanien|dollar du guyana|dollar dre guyana|$ guayana|gyd|$gy"],["Dollar de Hong Kong", "dollar hong kong|dollar hongkong|dollar de hong kong|dollar de hongkong|$hk|$ hk|hkd|hk $|hk$|dollar hk|$hongkong|dollars hongkong|dollars hong kong"],["Dollar jamaïcain", "dollar jamaïcain|dollars jamaïcain|dollar jamaicain|dollars jamaicain|$j|$ jamaïque|dollar jamaïque|jmd"],["Dollar libérien", "dollar libérien|dollars libérien|dollar liberien|dollars liberien|lrd|$ libérien|$ liberia|$ liberien"],["Dollar namibien", "dollar namibien|dollars namibien|$ namibien|nad|$n|dollar namibie|dollars namibie|$ namibie"],["Dollar des îles salomon", "dollar des îles Salomon|dollar des iles salomon|$si|sbd|$ iles salomon|$ îles salomon"],["Dollar du suriname", "dollar du suriname|srd|$ du suriname|$ suriname|dollar suriname|dollars suriname|dollars du suriname"],["Nouveau dollar de Taïwan", "nouveau dollar de taïwan|nouveau dollar de taiwan|twd|ntd|$nt"],["Dollar trinidadien", "dollar trinidadien|dollars trinidadien|ttd|$ trinidadien"],["Dollar", "dollar|$|dollars"],["Yuan Chinois", "yuan|yuans|yuan chinois|renminbi|cny|rmb|￥"],["Fen", "fen"],["Jiao", "jiao"],["Mark Finlandais", "marks finlandais|mark finlandais|fim|mark"]]);
-	export const CompoundUnitRegex = `(?<spacer>[^.])`;
+	export const CompoundUnitConnectorRegex = `(?<spacer>[^.])`;
 	export const CurrencyPrefixList: ReadonlyMap<string, string> = new Map<string, string>([["Dollar", "$"],["Dollar États-Unis", "$us|usd|us$"],["Dollar Caraïbes orientales", "xcd|$ec"],["Dollar australien", "a$|$a|aud"],["Dollar bahamas", "bsd|b$"],["Dollar barbadien", "bbd|bds$"],["Dollar de belize", "bz$|bzd"],["Dollar des bermudes", "bd$|bmd"],["Dollar de brunei", "brunei $|bnd"],["Dollar de Singapour", "s$|sgd"],["Dollar Canadien", "cad|$ ca|$ca|$ c"],["Dollar des îles Caïmans", "ci$|kyd"],["Dollar néo-zélandais", "nz$|nzd"],["Dollar de Fidji", "$fj|fjd"],["Dolar guyanien", "g$|gyd"],["Dollar de Hong Kong", "hkd|hk$"],["Dollar jamaïcain", "j$|jmd"],["Dollar libérien", "lrd|l$"],["Dollar namibien", "nad|n$"],["Dollar des îles Salomon", "$ si|$si|sbd"],["Nouveau dollar de Taïwan", "nt$|twd"],["Réal brésilien", "r$|brl|reais"],["Guaraní paraguayen", "₲|gs.|pyg"],["Dollar trinidadien", "ttd|titis"],["Yuan renminbi", "cny|rmb|¥|元"],["Yen", "¥|jpy"],["Euro", "€|eur"],["Pound", "£"],["Florín", "ƒ"],["Livre", "£|gbp"]]);
 	export const AmbiguousCurrencyUnitList = [ 'din.','kina','lari','taka','tala','vatu','yuan','bob','btn','cop','cup','dop','gip','jod','kgs','lak','mga','mop','nad','omr','sar','sbd','scr','sdg','sek','sos','std','try','yer' ];
 	export const InformationSuffixList: ReadonlyMap<string, string> = new Map<string, string>([["Bit", "-bit|bit|bits"],["Kilobit", "kilobit|kilobits|kb|kbit|kbits"],["Megabit", "megabit|megabits|Mb|Mbit|mégabit|mégabits"],["Gigabit", "gigabit|gigabits|Gb|Gbit"],["Terabit", "terabit|terabits|Tb|Tbit|térabit|térabits"],["Petabit", "petabit|petabits|Pb|Pbit|pétabit|pétabits"],["octet", "octet|octets|-octet"],["Kilooctet", "kilo-octet|kilo-octets|kilooctet|kilooctets|ko|kio|kB|KiB|kilobyte|kilobytes"],["Mégaoctet", "mégaoctet|mégaoctets|méga-octet|méga-octets|Mo|Mio|MB|mégabyte|mégabytes"],["Gigaoctet", "gigaoctet|gigaoctets|Go|Gio|GB|GiB|gigabyte|gigabytes"],["Téraoctet", "téraoctet|téraoctets|To|Tio|TB|TiB|térabyte|térabytes"],["Pétaoctet", "pétaoctet|pétaoctets|Po|Pio|PB|PiB|pétabyte|pétabytes"]]);
