+++ Patterns/Chinese/Chinese-NumbersWithUnit.yaml
@@ -310,7 +310,7 @@ CurrencySuffixList: !dictionary
 #Additional
     Finnish markka: 芬兰马克
     Penni: 盆尼
-CompoundUnitRegex: !simpleRegex
+CompoundUnitConnectorRegex: !simpleRegex
     def: (?<spacer>又|再)
 CurrencyPrefixList: !dictionary
   types: [string, string]
