+++ Patterns/German/German-NumbersWithUnit.yaml
@@ -369,7 +369,7 @@ CurrencySuffixList: !dictionary
 #Additional
     Finnish markka: suomen markka|finnish markka|finsk mark|fim|markkaa|markka|finnische mark|finnischen mark
     Penni: penniä|penni
-CompoundUnitRegex: !simpleRegex
+CompoundUnitConnectorRegex: !simpleRegex
     def: (?<spacer>[^.])
 CurrencyPrefixList: !dictionary
   types: [ string, string ]
