+++ Patterns/Portuguese/Portuguese-NumbersWithUnit.yaml
@@ -504,7 +504,7 @@ CurrencySuffixList: !dictionary
 #Zâmbia
     Kwacha zambiano: kwacha zambiano|kwacha zambianos|kwachas zambianos|zmw
     Ngwee zambiano: ngwee zambiano|ngwee zambianos|ngwees zambianos
-CompoundUnitRegex: !simpleRegex
+CompoundUnitConnectorRegex: !simpleRegex
     def: (?<spacer>[^.])
 CurrencyPrefixList: !dictionary
   types: [ string, string ]
