+++ Patterns/Spanish/Spanish-NumbersWithUnit.yaml
@@ -580,7 +580,7 @@ CurrencySuffixList: !dictionary
     Kwacha zambiano: kwacha zambiano|kwacha zambianos|zmw
     Ngwee zambiano: ngwee zambiano|ngwee zambianos
     #Zimbabue -> Dólar Estadounidense
-CompoundUnitRegex: !simpleRegex
+CompoundUnitConnectorRegex: !simpleRegex
     def: (?<spacer>[^.])
 CurrencyPrefixList: !dictionary
   types: [ string, string ]
